feat(app): add query operators example
we create queries from the command line to our database by using 'command-line-args' library, there

we specify 'commandLineOptions' method to see fields that we want to query and the user types its

values, that returns an 'options' object with values of the fields the user typed. There's other

method called 'queryDocument' that actually creates our query document with the values specified

from 'commandLineOptions' and that's use to retrieve the data we need.